requirements.txt
asgiref==3.6.0
babel==2.12.1
defusedxml==0.7.1
diff-match-patch==20230430
django-import-export==3.2.0
django-jazzmin==2.6.0
django-phonenumber-field[phonenumbers]==7.1.0
django==4.2.1
et-xmlfile==1.1.0
markuppy==1.14
odfpy==1.4.1
openpyxl==3.1.2
sqlparse==0.4.4
tablib[html,ods,xls,xlsx,yaml]==3.4.0
tzdata==2023.3
xlrd==2.0.1
xlwt==1.3.0
psycopg2==2.9.6
django-colorfield==0.8.0
pytz==2023.3
pytelegrambotapi==4.7.1
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.13
